many changes.
No longer require escaping functions for content / labels (but does it make sense to rewrap content later?)

Configuration now works over a generic selection type, even though that selection type always happens to be AnyHashable. The style protocol currently does this as a stand-in for proper types. It might make sense to use an intermediate step somewhere higher up in the chain in order to erase the type to allow the innards to be more generic as well.